add environment change hook
