+++ crates/nu-protocol/src/engine/mod.rs
@@ -2,10 +2,12 @@ mod call_info;
 mod capture_block;
 mod command;
 mod engine_state;
+mod overlay;
 mod stack;
 
 pub use call_info::*;
 pub use capture_block::*;
 pub use command::*;
 pub use engine_state::*;
+pub use overlay::*;
 pub use stack::*;
