+++ docs/sample_config/default_config.nu
@@ -206,6 +206,11 @@ let-env config = {
     pre_execution: [{
       $nothing  # replace with source code to run before the repl input is run
     }]
+    env_change: {
+      PWD: [{|before, after|
+        $nothing  # replace with source code to run if the PWD environment is different since the last repl input
+      }]
+    }
   }
   menus: [
       # Configuration for default nushell menus
